feat: added pending-rewards module
